check if installs are needed

+++ install.sh
@@ -2,7 +2,7 @@
 
 echo "---------- Installer script for dotfiles ----------"
 echo "asking for sudo access"
-sudo su
+sudo -v
 
 # Backups and copying files
 mkdir "backup_dotfiles"
@@ -16,9 +16,10 @@ fi
 echo "checking for vim plugins..."
 if [ ! -e $HOME/.vim/pack/plugins/start/lightline ]; then
     git clone 'https://github.com/itchyny/lightline.vim.git' "$HOME/.vim/pack/plugins/start/lightline"
-
+fi
 if [ ! -e $HOME/.vim/pack/plugins/start/lightline ]; then
     git clone 'https://github.com/frazrepo/vim-rainbow.git' "$HOME/.vim/pack/plugins/start/rainbow"
+fi
 
 # Blocking ads and malware hosts
 echo "blocking ads and malware"
@@ -47,23 +48,26 @@ else
     echo "installation failed, apt or pacman not available"
 fi
 
-echo "installing oh-my-zsh"
-read -n1 -p "type 'exit' to get back to install process after oh-my-zsh screen appears. \nPress any key to continue..." tmp
-sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-# echo "source ~/.zshrc.pre-oh-my-zsh" >> ~/.zshrc
+echo "checking for oh-my-zsh"
+if [ ! -e $HOME/.oh-my-zsh ]; then
+    read -n1 -p "type 'exit' to get back to install process after oh-my-zsh screen appears. \nPress any key to continue..." tmp
+    sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+    # echo "source ~/.zshrc.pre-oh-my-zsh" >> ~/.zshrc
 
-echo "copying oh-my-zsh plugins"
-git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+    echo "copying oh-my-zsh plugins"
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+fi
 
-echo "copying files"  # copying files after all installs to ensure they aren't overwritten by install processes
 echo "replacing config files"
 cp -r ".zshrc" ".vimrc" ".vim" ".p10k.zsh" ".config" "$HOME"
 
 # Configuring shells
-echo 'Changing preferred shell to zsh'
-chsh -s /bin/zsh
+if [[ $SHELL == "/bin/bash" ]]; then
+    echo 'Changing preferred shell to zsh'
+    chsh -s /bin/zsh
+fi
 
 if  which conda ; then
     echo initialising conda
@@ -77,16 +81,19 @@ else
 fi
 
 # Git configuration
-read -n1 -p "Setup git config: [y]/n" configure_git
-if [ $configure_git != n ] ; then
-    read -p "\nname: " git_name
-    read -p "email: " git_email
-    echo "# This is Git's per-user configuration file.
-[user]
-# Please adapt and uncomment the following lines:
-    name = $git_name
-    email = $git_email
-" > $HOME/.gitconfig
+if [ ! -e $HOME/.gitconfig ]; then
+    read -n1 -p "Setup git config: [y]/n" configure_git
+    if [ $configure_git != n ]; then
+        read -p "name: " git_name
+        read -p "email: " git_email
+		cat <<END > $HOME/.gitconfig
+# This is Git's per-user configuration file.
+    [user]
+    # Please adapt and uncomment the following lines:
+        name = $git_name
+        email = $git_email
+END
+    fi
 fi
 
-echo "\n---------- Install finished ----------"
+echo "---------- Install finished ----------"
